contenido/cloud/code_pattern-ejecutar_tareas_periodicas_en_dos_simples_pasos/source/insert-current-time-mongodb/requirements.txt
pymongo
python-dotenv
